created header and section components
